+++ include/aligator/gar/dense-riccati.hpp
@@ -1,8 +1,6 @@
-/// @copyright Copyright (C) 2024 LAAS-CNRS, INRIA
+/// @copyright Copyright (C) 2024 LAAS-CNRS, 2024-2025 INRIA
 /// @author Wilson Jallet
 #pragma once
-
-#include <proxsuite-nlp/linalg/bunchkaufman.hpp>
 
 #include "riccati-base.hpp"
 #include "dense-kernel.hpp"